improved listner interface + receiver actor
