+++ src/main/java/org/zalando/nakadi/client/EventListener2.java
@@ -1,6 +1,6 @@
 package org.zalando.nakadi.client;
 
-public interface EventListener2 extends EventListener {
+public interface EventListener2 extends Listener {
     void onConnectionClosed(final String topic, final String partitionId);
     void onConnectionClosed(final String topic, final String partitionId, final Exception cause);
 }
