+++ src/main/scala/org/zalando/nakadi/client/Listener.scala
@@ -1,0 +1,7 @@
+package org.zalando.nakadi.client
+
+
+trait Listener {
+  def onReceive(topic: String, partition: String, cursor: Cursor, event: Event)
+  def onConnectionClosed(topic: String, partition: String, lastCursor: Cursor)
+}
